# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API du registre parcellaire de Kinshasa
 * OpenAPI spec version: 0.1.0
 */
export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin_principal: 'admin_principal',
  validateur: 'validateur',
  agent: 'agent',
} as const;

export interface AuthUser {
  id: string;
  email: string;
  role: UserRole;
}

export interface ManagedUser {
  id: string;
  email: string;
  role: UserRole;
  active: boolean;
  createdAt: string;
}

export interface LoginInput {
  /** @maxLength 254 */
  email: string;
  /**
     * @minLength 20
     * @maxLength 256
     */
  code: string;
}

export interface UserInput {
  /** @maxLength 254 */
  email: string;
  role: UserRole;
}

export interface UserUpdate {
  role?: UserRole;
  active?: boolean;
}

export interface CreatedUser {
  user: ManagedUser;
  accessCode: string;
}

export interface AccessCode {
  accessCode: string;
}

export interface LogoutResult {
  success: boolean;
}

export interface HealthStatus {
  status: string;
}

export interface Commune {
  code: string;
  nom: string;
}

export type FicheHygiene = { [key: string]: unknown };

export type FicheDechets = { [key: string]: unknown };

export type FicheFacade = { [key: string]: unknown };

export type FicheDrainage = { [key: string]: unknown };

export type FicheAvis = { [key: string]: unknown };

export type FicheStatutFiche = typeof FicheStatutFiche[keyof typeof FicheStatutFiche];


export const FicheStatutFiche = {
  brouillon: 'brouillon',
  soumise: 'soumise',
  validee: 'validee',
  rejetee: 'rejetee',
} as const;

export type FicheStatutPlaque = typeof FicheStatutPlaque[keyof typeof FicheStatutPlaque];


export const FicheStatutPlaque = {
  non_generee: 'non_generee',
  generee: 'generee',
  a_reimprimer: 'a_reimprimer',
  imprimee: 'imprimee',
} as const;

export interface Fiche {
  id: string;
  ficheNo: string;
  commune: string;
  quartier: string;
  /** @nullable */
  localite: string | null;
  avenue: string;
  parcelleNo: string;
  /** @nullable */
  plaqueNo?: string | null;
  proprietaireNom: string;
  telephone: string;
  typeOccupation: string;
  /** @nullable */
  superficie?: number | null;
  usageParcelle: string;
  /** @nullable */
  plaqueExistante?: string | null;
  /** @nullable */
  statutPaiement?: string | null;
  /** @nullable */
  recuNo?: string | null;
  /** @nullable */
  sensibilisation?: string | null;
  hygiene?: FicheHygiene;
  dechets?: FicheDechets;
  facade?: FicheFacade;
  drainage?: FicheDrainage;
  activites?: string[];
  /** @nullable */
  remarques?: string | null;
  avis?: FicheAvis;
  /** @nullable */
  agentMatricule?: string | null;
  /** @nullable */
  chefRueNom?: string | null;
  /** @nullable */
  chefRueAvenue?: string | null;
  statutFiche: FicheStatutFiche;
  statutPlaque: FicheStatutPlaque;
  dateProspection: string;
  createdAt: string;
}

export type FicheInputHygiene = { [key: string]: unknown };

export type FicheInputDechets = { [key: string]: unknown };

export type FicheInputFacade = { [key: string]: unknown };

export type FicheInputDrainage = { [key: string]: unknown };

export type FicheInputAvis = { [key: string]: unknown };

export interface FicheInput {
  /** @minLength 1 */
  commune: string;
  /** @minLength 1 */
  quartier: string;
  /**
     * @maxLength 120
     * @nullable
     */
  localite?: string | null;
  /** @minLength 1 */
  avenue: string;
  /** @minLength 1 */
  parcelleNo: string;
  plaqueNo?: string;
  /** @minLength 1 */
  proprietaireNom: string;
  /** @minLength 1 */
  telephone: string;
  typeOccupation: string;
  superficie?: number;
  usageParcelle: string;
  plaqueExistante?: string;
  statutPaiement?: string;
  recuNo?: string;
  sensibilisation?: string;
  hygiene?: FicheInputHygiene;
  dechets?: FicheInputDechets;
  facade?: FicheInputFacade;
  drainage?: FicheInputDrainage;
  activites?: string[];
  remarques?: string;
  avis?: FicheInputAvis;
  agentMatricule?: string;
  chefRueNom?: string;
  chefRueAvenue?: string;
  dateProspection: string;
}

export type FicheUpdate = FicheInput;

export interface FicheLocaliteInput {
  /**
     * @maxLength 120
     * @nullable
     */
  localite: string | null;
}

export type FicheDecisionDecision = typeof FicheDecisionDecision[keyof typeof FicheDecisionDecision];


export const FicheDecisionDecision = {
  validee: 'validee',
  rejetee: 'rejetee',
} as const;

export interface FicheDecision {
  decision: FicheDecisionDecision;
  motif?: string;
}

export interface DuplicateCheck {
  duplicate: boolean;
  fiche?: Fiche | null;
}

export type PlaqueStatut = typeof PlaqueStatut[keyof typeof PlaqueStatut];


export const PlaqueStatut = {
  generee: 'generee',
  a_reimprimer: 'a_reimprimer',
  imprimee: 'imprimee',
} as const;

export interface Plaque {
  id: string;
  ficheId: string;
  ficheNo: string;
  commune: string;
  quartier: string;
  avenue: string;
  plaqueNo: string;
  version: number;
  statut: PlaqueStatut;
  svg: string;
  genereLe: string;
  /** @nullable */
  imprimeLe?: string | null;
}

export interface RubriqueSettings {
  adressage: boolean;
  hygiene: boolean;
  dechets: boolean;
  facade: boolean;
  drainage: boolean;
  activites: boolean;
  remarques: boolean;
  avis: boolean;
}

export type RubriqueSettingChangeRubrique = typeof RubriqueSettingChangeRubrique[keyof typeof RubriqueSettingChangeRubrique];


export const RubriqueSettingChangeRubrique = {
  adressage: 'adressage',
  hygiene: 'hygiene',
  dechets: 'dechets',
  facade: 'facade',
  drainage: 'drainage',
  activites: 'activites',
  remarques: 'remarques',
  avis: 'avis',
} as const;

export interface RubriqueSettingChange {
  rubrique: RubriqueSettingChangeRubrique;
  active: boolean;
}

export type DashboardRepartitionCommunesItem = {
  commune: string;
  total: number;
};

export type DashboardActiviteRecenteItem = {
  id: string;
  action: string;
  ficheNo: string;
  horodatage: string;
};

export interface Dashboard {
  totalFiches: number;
  enAttente: number;
  validees: number;
  plaquesAImprimer: number;
  repartitionCommunes: DashboardRepartitionCommunesItem[];
  activiteRecente: DashboardActiviteRecenteItem[];
}

export type GetDashboardParams = {
commune?: string;
};

export type ListFichesParams = {
commune?: string;
statut?: string;
recherche?: string;
};

export type CheckFicheDuplicateParams = {
commune: string;
quartier: string;
avenue: string;
parcelleNo: string;
};

export type ListPlaquesParams = {
commune?: string;
statut?: string;
};