# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API du registre parcellaire de Kinshasa
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface AuthUser {
  id: string;
  email: string;
  role: UserRole;
}