# lib/api-zod/src/generated/types/createdUser.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API du registre parcellaire de Kinshasa
 * OpenAPI spec version: 0.1.0
 */
import type { ManagedUser } from './managedUser';

export interface CreatedUser {
  user: ManagedUser;
  accessCode: string;
}