# lib/api-zod/src/generated/types/loginInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API du registre parcellaire de Kinshasa
 * OpenAPI spec version: 0.1.0
 */

export interface LoginInput {
  /** @maxLength 254 */
  email: string;
  /**
     * @minLength 20
     * @maxLength 256
     */
  code: string;
}