# lib/api-zod/src/generated/types/managedUser.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API du registre parcellaire de Kinshasa
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface ManagedUser {
  id: string;
  email: string;
  role: UserRole;
  active: boolean;
  createdAt: Date;
}