# lib/api-zod/src/generated/types/userInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API du registre parcellaire de Kinshasa
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface UserInput {
  /** @maxLength 254 */
  email: string;
  role: UserRole;
}