# lib/api-zod/src/generated/types/userRole.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API du registre parcellaire de Kinshasa
 * OpenAPI spec version: 0.1.0
 */

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin_principal: 'admin_principal',
  validateur: 'validateur',
  agent: 'agent',
} as const;